[OPIK-7105] [FE] refactor: colocate the pairing URL helper with its guard
getWorkspaceNameFromUrl was parked in lib/utils.ts when it was rescued from the
deleted lib/workspaceVersion.ts. It has exactly one consumer,
PairRouteWorkspaceGuard, so it is not a repo-wide utility.

.agents/skills/opik-frontend/code-quality.md is explicit: helpers shared by a
component family belong in a module next to them, not in lib/ -- "lib/ is for
repo-wide utilities". Moves the function and its private path-segment helper
into the guard's own folder. Behaviour is unchanged.

Raised in review.
